requirements.txt
fastapi[standard]==0.115.0
uvicorn[standard]==0.30.6
python-dotenv
elasticsearch
ibm-watsonx-ai